Rename wsi::Window to wsi::PlatformWindow
